Provide better file pointers for defective override files. (Closes: #1000494)
The previous format made it harder to figure out which file was affected. Here
we use the brand-new file pointer facility.

Sample hints are below.

Please note that overrides for installable packages must be checked against the
installed override file. That is not the location a maintainer is used to, but
the source package, which also ships the file, does not trigger the hints.

Thanks to Marc Haber for bringing the matter to our attention!

* * *

➤ bin/lintian /mirror/debian/pool/main/m/mingw-ocaml/mingw-ocaml_4.01.0\~20140328-1_all.deb
E: mingw-ocaml: malformed-override Expected a colon [usr/share/lintian/overrides/mingw-ocaml:1]
E: mingw-ocaml: malformed-override Expected a colon [usr/share/lintian/overrides/mingw-ocaml:2]