For manual references, download the vim package instead of looking in the file system.
The links detected in the vim-doc versions in stable is significantly different
from the version in unstable. (For details, please see Bug#1001247.) For now, we
continue to ship the version in stable.

The rationale for stable is that it is the preferred and, well, more stable
system for many users. It is also what one of the Lintian maintainer uses.

A counter argument would be that the share of folks on unstable is greater
among Debian contributors, who are more likely to rely on citations of authority
like the Vim packaging policy.

No longer requires the vim package to be installed locally.

Gbp-Dch: ignore