Show numeric sections in Markdown for FHS citations; refresh FHS data.
The FHS website is unique because, in the tag files, we use their word anchors
to point to sections but those headings do not show well, especially not in all
lowercase letters.

The website actually uses camel-case.

Here we also collect the traditional, numeric section keys and show those to
users instead.

Gbp-Dch: ignore