Provide rudimentary Emacs integration. (See: #968758)
Provides the 'lintian-run-file' command to view packaging hints in Emacs. If a
user's teminal cannot display OSC8 hyperlinks [1] it is a way to get live links
to tag explanations.

Unfortunately, the simple shr renderer used here does not process CSS. In many
ways, the output is less than ideal, but it's a start!

Please activate the functionality with '(load "lintian.el")' in your init.el.

[1] https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda