Exempt installables designated as documentation from warning about new Python2 packages. (Closes: #995286)
The check already meant to skip documentation packages, but the code to detect
the condition had two issues.

First, parentheses were missing around the 'none' condition. Second, the quoted
pattern required a dollar sign in the installable name for the exemption to be
effective. Both are fixed here.

We kept the existing logic based on the installable name but could easily
switch to 'Section: doc' idea (which I personally like better). Please just us
know.

The existing logic is used in four other places in the same check.

The bug report did not refer to a particular case (and the website showed only
seven false positives) so we checked with:

➤ bin/lintian /mirror/debian/pool/main/n/network-runner/python-network-runner-doc_0.2.2-1_all.deb
I: python-network-runner-doc: capitalization-error-in-description python Python
I: python-network-runner-doc: possible-documentation-but-no-doc-base-registration

The offending hint is now gone.

Thanks to Andrius Merkys for bringing the matter to our attention!