Drop version requirement from skip-systemd-native-flag-missing-pre-depends. (See: #1003271)
The tag combines two checks in one. First, it requires that a Pre-Depends on
was declared and, second, that the version is sufficient. The reporting party
pointed out that the version requirement is satisfied in oldstable, so the
version requirement was dropped from both the code and from the description.

The overall check that a Pre-Depends was declared, however, may still be
sensible to keep.

On the website 73 sources currently provoke the tag, but we cannot if the
declaration is missing or just does not satisfy the minimum version requirement.
Let's see how the tag performs after this change.

Thanks to Tobias Frost for bringing the matter to our attention!