Recognize dh-sequence-sphinxdoc as a valid prerequisite for dh_sphinxdoc. (Closes: #999810)
This commit does not yet address the related issue in the cloned Bug#1000234,
which belongs to a larger family of merged bugs.

Moreover, this fix makes it impossible to check going forward that the upcoming
commit for the cloned bug actually solves it. I verified it separately before
applying this commit.

The sources mentioned in the bug report now show an unused override:

➤ bin/lintian -X cruft /mirror/debian/pool/main/c/cmake/cmake_3.22.1-1.dsc
I: cmake source: unused-override missing-build-dependency-for-dh_-command dh_sphinxdoc => * [debian/source/lintian-overrides:8]
[unrelated output omitted]