Require the targets build-arch and build-indep in d/rules. (Closes: #657390)
At the time of writing, 264 sources were still affected, down from 421 in
November. That worked out more or less to the monthly decline of 85 sources
originally predicted by N. Thykier in 2013.

Lintian's tag description was amended with the simplest possible fix, which has
two-lines. They are a reasonable addition to sources that do not receive a lot
of attention.

For more information, please check the thread on debian-devel in the tag
references.