Exempt the names of Debian folks associated with a package from spelling checks. (Closes: #989381)
Sorry about the delay in resolving this matter. We exempted the names of all
persons mentioned in the relevant Debian fields.

The offending hint is now gone:

➤ bin/lintian /mirror/debian/pool/main/p/python-cdsapi/python3-cdsapi_0.5.1-1_all.deb
[no output]

Another improvement might be to add copyright holders from a DEP-5 copyright
file, if one is available.

Thank you to Gard Spreemann for bringing the matter to our attention, and also
for being so good-natured and lighthearted about that very personal matter!