Prevent issuance of tags under their old names.
Julian Gilbey was able to spot [1] a super obscure bug introduced recently when
the handling of tags was revamped so that old tag names could be shown on the
website along with their current names.

[1] https://salsa.debian.org/lintian/lintian/-/commit/5e20ecb49725912c7d7a42a721a38987981fd0dd#note_291378

The bug was probably introduced in commit 3ad636e37a1.

Lintian now complains, as it should, about the old tag names being issued with
the two warnings below. This commit also fixes those warnings.

Warning in processable [...]/requests-undeclared-python-versions_1.0.dsc:
 Tried to issue renamed tag declare-requested-python-versions-for-test
 (current name declare-python-versions-for-test)
 in check testsuite. at [...]/lib/Lintian/Group.pm line 333.

Warning in processable [...]/does-not-query-declared-python-versions_1.0.dsc:
 Tried to issue renamed tag query-requested-python-versions-in-test
 (current name query-declared-python-versions-in-test)
 in check testsuite. at [...]/lib/Lintian/Group.pm line 333.

Thanks to Julian for his extraordinary attention!

Gbp-Dch: ignore