Flag an outdated Debian copyright just once; use the most recent year. (Closes: #1003817)
Picks just the most recent of all Debian years now. Points to the topmost
section, if that year appears more than once. By convention, that should be the
broadest pattern and therefore the most likely in need of an update.

Four existing test cases were modified to validate the desired behavior going
forward.

Thanks to Mattia Rizzolo for bringing the matter to our attention!