Deprecate --no-tag-display-limit for '--tag-display-limit 0'; update documentation. (Closes: #1004239)
> tag-display-limit can be set from the configuration file
> this is not mentioned in the manual page.

Fixed here

> would it not make more sense to make all command line options configurable
> from the configuration file and only mention the exceptions?

We love suggestions on how to make Lintian simpler and more intuitive, but
there are 42 command-line options [1] and just sixteen configuration options.
[2] Your ideas are welcome!

I have been somewhat stingy with configuration option because we may have to
support them forever.

[1] https://salsa.debian.org/lintian/lintian/-/blob/master/bin/lintian#L136-177
[2] https://salsa.debian.org/lintian/lintian/-/blob/master/bin/lintian#L259-276

> would it make sense to have tag-display-limit=0 a synonym
> to no-tag-display-limit?

It has been a synonym for a while, but only on the command-line, where
'--no-tag-display-limit' is now deprecated in favor of '--tag-display-limit 0'.

In configuration files, the setting 'no-tag-display-value' was previously not
supported.

> It seems surprising [...] that no-tag-display-limit cannot be set in the
> configuration file.

Please use 'tag-display-limit=0' in your configuration file.

Thanks to Marc Haber for bringing all these matters to our attention!