Mask examples in tests from package-does-not-install-examples. (Closes: #1005184)
The sources mentioned now produce this output:

➤ bin/lintian /mirror/debian/pool/main/p/python-tomlkit/python-tomlkit_0.8.0-1.dsc /mirror/debian/pool/main/p/python-tomlkit/*_0.8.0-1*.deb
N: masked by screen examples/in-tests
M: python-tomlkit source: package-does-not-install-examples [tests/examples/]
[unrelated output omitted]

Please note that this tag is sometimes hard to reproduce. It only shows when
both the sources and the installables are present on the command line.

Thank you to Scott Kitterman for bringing the matter to our attention!