Silence a very widespread false positive for detached debug symbols
binutils reports that it is unable to find the program interpreter name
for detached debug symbols (#1000449), but it's entirely reasonable
that detached debug symbols don't have an interpreter, since they are
not a complete executable program.

This false positive is so widespread that it's making Lintian less
useful for detecting other issues, and it seems like this is not going
to be solved in binutils any time soon, so silence it to improve the
signal-to-noise ratio.

Closes: #1000977