Store ELF information from readelf in an MLDBM database. (Closes: #1003456)
Works but is slow. Perhaps not quite so much of the readelf information needs
to be retained.

The impact on average library packages has yet to be evaluated.

It may only is be issue for some installables. The one from the bug, for
example (picolibc-arm-none-eabi_1.7.4-1_all.deb) seemed to generate several
million symbols whose purpose was not clear to me. That impression is dated
because I looked into the reason for the high memory usage in January.