Upgrade missing-systemd-timer-for-cron-script to warning; no longer experimental. (Closes: #1007257)
Until now, the visibility was experimental and pedantic, which is the lowest
combination still shown to users. The requested change effectively bumps the
visibility by three levels.

The change is more dramatic than what we would normally entertain in a single
step.

At the same time, the tag has existed for over three years. (See Bug#922862.)
In addition, the website showed only 234 hits in 185 sources across the entire
archive.

While some users may disagree with the change, the risk of widespread
irritation is low.

Thanks to Josh Triplett for bringing the matter to our attention!