init.d-general check: Avoid relying on line numbers in #DEBHELPER# replacement code
We add our own second update-rc.d call for that. The replacement for
the #DEBHELPER# token will add a third one, but lintian will only
argue about the first such duplicate. (Which might be a minor issue
itself, but is for now our luck here. :-)

Additionally add "> /dev/null" to the first update-rc.d call to avoid
triggering the output-of-updaterc.d-not-redirected-to-dev-null.

This is currently not needed for the second call, which actually seems
to be another potential bug, see https://bugs.debian.org/1012732

This will hopefully be the last commit which was needed to make
Lintian's testsuite to pass again. :-)