Add unit test for Lintian::Util::match_glob
Also fix the anchoring of regular expressions returned by
Regexp::Wildcards. The test showed that it wasn't implemented as
expected (i.e. as before Text::Glob was used). It should do just
substring matches but match the whole line (if patterns are present).

This is basically meant as a test case for #1003353 because writing a
test that tests if lintian overrides actually match seems to be
impossible with the current way Lintian's test suite works.

The weird indentation in the test script is (C) by perltidy. O.o