Fix "Use of uninitialized value $step" in Lintian::Version
This only shows up if running the test suite against a checkout git
repo where the latest commit is a tagged lintian repository, i.e. "git
describe" doesn't contain any dashes in the output.

Yes, this is a last-minute fix and will have to do some rebasing to
make it show up in git history before the final changelog touching for
the release. :-)