run-private-scripts.t: Skip some not shipped scripts under autopkgtest
I want all these scripts to checked for their usefulness because of
what happened to previously untested latest-policy-version script. (I
was causing autopkgtest failures in libconfig-model-dpkg-perl.) This
is even more because that script is currently not run when building or
using lintian.

But some of the other scripts are not going to be installed into the
binary package and since autopkgtest runs the same testsuite also on
the installed binaries we should skip them under autopkgtest, but
still should be checked when running the test suite locally.

Also fix one case in the same script where double quotes were meant to
be used but weren't.

Gbp-Dch: Ignore