uses-pdm-cli: create new tag.
This new tag verifies people building using the python pdm PEP517-style
backend use the python3-pdm-pep517 (used to build packages) instead of
python3-pdm (the cli tool used to create python projects with pdm).

This tag and the tests it comes with are heavily inspired by the
"uses-poetry-cli" tag, that checks for a similar thing.