Add pedantic hint for OpenPGP files named after specific implementations
The specification for this file format is called OpenPGP, and the
extension name that is short and considered implementation neutral
is ".pgp".

While currently the GnuPG project is widely used and one of the most
known OpenPGP implementations, using an extension after its name is
detrimental to other alternative implementations, when a better more
neutral name can be used instead (".pgp").

Signed-off-by: Guillem Jover <guillem@debian.org>