Fix --fail-on to revert to original default on error
There was a request to make it possible to exit 0 regardless of error
hints being emitted, which can be useful for systems that want to report
all hints. The request got implemented and made the default. The problem
is that this behavior change implies any automated system, infrastructure,
script, or similar relying on the exit code reporting a non-zero code for
an error hint, has been ignoring them since this change, and would require
amending all call sites to pass --fail-on=error to restore the previous
behavior. This is compounded with the behavior being silent or invisible
to automated systems, and then letting through clearly bogus packages.

A patch was provided in #962158, but it got partially applied, which
meant that the old behavior never got properly reverted. This change
fixes that.

Restore the previous default to fail on hint errors, so that users and
CI systems can detect that there was an error from the exit code,
without needing to update all their scripts calling lintian, directly or
indirectly.

Fixes: commit 3758bfafd5dd742c327f2312dac8e3a71b1f036e
Closes: #995492