Fix arm64 autopkgtest by removing the bin-sbin-confusion-in-elf test from the test suite
The test bin-sbin-confusion-in-elf in Lintian's test suite is broken
beyond repair as it currently expects the false negatives it has on at
least amd64, armhf and i386 as string constants optimization removes
some literal trigger keywords from the binary at compile time.

The arm64 results, which failed the test actually showed all results
it should show, but the test suite expected the incomplete result seen
on archictures with this optimization.

So in the end this test can't really check if Lintian is right or
wrong. So it's mostly useless.

Closes: #1025868

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1025868#44 for a
detailed analysis on this issue.