[Testsuite] Fix armhf+i386 only test binaries-missing-lfs
It seems to have always failed since the migration to pointed hints,
but nobody noticed so far as failing autopkgtest on these
architectures is not considered to be release-critical.

This should fix autopkgtest on armhf and i386. It probably also fixes
FTBFS on i386 and armhf due to test suite failures, at least since the
test suite is run at build time again.