field-too-long: allow X-Cargo-Built-Using
this field is emitted by dh-cargo-built-using, the length of the field
is directly related to the size of the (transitive) (build-)dependency
tree.

the field needs to capture the complete (transitive build-)dependency
tree in order to determine which packages need to be rebuilt (e.g.
during freeze time to ensure no outdated builds are part of a
soon-to-be-released Debian version, or after some library crate was
patched to fix a security issue or other serious bug), since rust
doesn't have (usable) dynamic linking yet, and all shipped rust binaries
are statically linked with their crate dependencies (regular
C-compatible libraries like libssl/.. are linked dynamically).

as an alternative, a second much higher limit for this and some other
fields might make sense as well, as originally discussed in #942493

Signed-off-by: Fabian Grünbichler <debian@fabian.gruenbichler.email>