Fix exit code when --show-overrides is given and an 'error' tag is overridden
Previously, if an error tag was overridden but --show-overrides was given,
lintian would correctly override the tag, but incorrectly give an exit
code of 2.

The reason is a logic error in lib/Lintian/Pool.pm in the code from lines 173-212

The whole block seems to be attempting to:
- skip hints (tags) that should not be displayed (experimental, or at a too-low display-level)
- count number of errors/warnings that remain (used in line 217 to set exit code)
- keep hints that should be displayed (line 211)

Because on line 194, if an 'error' tag is overridden and --show-overrides is given the block
decided not to call next (because keeping the hint doesnt happen until line 211).

But then line 197 is executed which means that the hint is counted as a warning/error/etc whenever
show-overrides was present: effectively show-overrides makes it look like no override was present.

(And then on line 217 the exit_code will be set to 2).

This patch therefore changes 197 to only run the sub-block if an override was not defined.