Only run the binaries/corrupted check on elf files
Under some rare conditions this check can hang for over 5 minutes on a
single file. This file can be a source code file containing just about
70 lines of code. Unfortunately, this is hard to reproduce. If you
create a debian package that only includes this file, the issue doesn't
occur. So, it seems to be related to some other conditions.
Nevertheless, this check does only make sense for elf files because it
tests elf issues. For that reason, just skip the check on non elf files.