Fix inconsistencies in documentation of lintian-explain-tags
The man page is inconsistent with the description of this tag:

In the man page synopsys it shows that it can be used:
SYNOPSIS
       lintian-explain-tags --tags tag ...

The OPTIONS section disagrees:
       -t, --tag, --tags
           This option has no effect. It exists for historical reasons.

And the --help message agrees with the synopsis:
    -t, --tag, --tags  display tag descriptions

Finally, using the tool indicates that the synopsis is correct:

$ lintian-explain-tags -t windows-devel-file-in-package
N:
W: windows-devel-file-in-package
N:
N:   This package appears to contain development files only meaningful to Windows environments. Such files are generally useless in Debian packages and were usually accidentally included by copying complete
N:   directories from the source tarball.
N:
N:   Visibility: warning
N:   Show-Always: no
N:   Check: foreign-operating-systems

This commit updates the OPTIONS section to agree with the rest.