gobject-introspection | dh-sequence-gir implements dh --with=gir
The debhelper addon is actually in gobject-introspection-bin, but that
package is an implementation detail and should not be used directly.

The use of :any here is actually incorrect, because gobject-introspection
is architecture-dependent (Multi-Arch: same) and we need specifically
the version for the host architecture, not the version for the build
architecture or any other version. However, Lintian does not currently
seem to recognise the dependency as being satisfied if we do not add
the :any suffix.

Closes: #964290, #1063709