ci: check longer timeout for autopkgtest
The default value of CI_JOB_TIMEOUT is 1h (3600), but autopkgtest will
not finish in 1 hour generally.

lintian/lintian CI/CD setting of timeout is extended to 3h, but
not true for forked/contributors lintian CI/CD settings - so it always
fail.

This MR aimed to fix warn such a situation in advance.

Signed-off-by: Kentaro Hayashi <kenhys@gmail.com>