Validate domain's TLD when checking for bogus-mail-host.
Closes: #1081316

The `domain_disable_tld_validation => 1` variable in
`Data::Validate::Domain` used not to do anything, since the version of
this library in Debian (0.10) had not implemented it. This feature has
only been available since 0.13. Upgrading the library to 0.15 added the
feature and thus broke out testsuite.

Since the goal of the check was to validate TLDs anyway, this patch restores
the previous behavior.