Make File::Contents::LineLength::visit_patched_files a bit faster.
No need to decode utf8 for lines shorter than our current longest
(since the utf8 for a bit of text is always more bytes than it has
codepoints)

No need to check sql insert/update statements on every line, only
the longest ones. And only need to check if we're in an sql file
a single time up front.