Indent some lists with a single space to fix console output
It seems lintian-explain-tags does not really understand the Markdown
list syntax, so to keep a single entry by line in the console output
we indent the lists with a single space. This will not impact the
output of Lintian SSG, as the list parser has a higher priority than
the code block parser. Also, most of the lists are already indented
this way.