Warn about dependency on obsolete bootstrap packages
twitter-bootstrap3 and twitter-bootstrap4 have been EOL'd upstream and
public security fixes are no longer available. Packages should migrate
to bootstrap 5 packages (src:bootstrap-html), but that requires work on
the upstream side. The new tags should help to make the maintainers
aware of the dependency on obsolete packages and where to find
migration guidelines.

Closes: #1086740