clarify nmu-in-changelog tag warning
I had this in my changelog:

 * Merge NMU

Which lintian thought I was mistakenly declaring my upload as a NMU. I
wasn't, but okay, I shouldn't call it a "merge". So I wrote this:

 * Acknwoledge NMU

And it STILL didn't work. WTF? LINTIAN WHAT AM I SUPPOSED TO DO?!

Did you spot the typo? Yes? Well, you're smarter or less tired than
me.

So let's help our users here and spell out something they can just
copy-paste.

This covers the case where the maintainer tries to recognize a
previous NMU upload in the commitlog but does so in a way that lintian
doesn't recognize as an acknowledgement. lintian then treats this as
the changelog announcing that it's a NMU, while it is not actually a
NMU, kind of the opposite.

This was the changelog that triggered the warning (and shouldn't:

    requests-file (2.1.0-1) unstable; urgency=medium

     * Acknwoledge NMU
     * New upstream release

And this is the fixed changelog that silenced the warning:

    requests-file (2.1.0-1) unstable; urgency=medium

     * Acknowledge NMU
     * New upstream release