Add support and checks for the Static-Built-Using field
The Static-Built-Using field was introduced from dpkg 1.21.3 for the
purpose of listing extra source packages that were used during the
build of a binary package. There is currently a proposal at #1069256 to
add this field to the Debian Policy as well.

This commit adds the Static-Built-Using field to the list of known
fields, and adds a check to ensure that the value of the field lists
source packages with an exactly equal ("=") version relation (same
format as Built-Using).

This commit also modifies the languages/golang/built-using check to
ensure that Static-Built-Using is used instead of Built-Using for
listing Golang libraries incorporated in a binary package.

Closes: #1050428