Update check for dependency on python3-numpy-abi
The release of NumPy 2 means that the name of the virtual package has
changed from python3-numpy-abiN to python3-numpy2-abiN. This commit
updates the check to account for this change.

This commit also updates the check to allow an alternative dependency
specifically on "python3-numpy2-abi0 | python3-numpy-abi9", since the
first ABI release of NumPy 2 is backwards compatible with the last ABI
release of NumPy 1, and this is what dh_numpy generates.

The ability to add a strict dependency on python3-numpy without
depending on the -abi package has been removed, since this practise is
obsolete and it's preferred to depend on the -abi package instead.

Closes: #1094364