Revert deletion of debian/source/local-options
This reverts the deletion of this file made in
b15eeaf4d7d1338305bb326b8b5ef88a453974fe

The NMU did contain this deletion and so was imported as-is, but the
debdiff in BTS #1096102 did not contain this deletion.

As such I'm taking for granted it was a mistake and should not have been
deleted.