Mention StaticLinking wiki page for go programs
The statically-linked-binary check explicitly excludes golang programs. At
first glance it may seem that binaries produced by the go compiler are actually
linked dynamically, but that is only partially true (only libc is in fact
dynamically linked), and only for programs not using any cgo. Point to the wiki
page explaining this.