Remove silent-on-rules-requires-root tag and tweak condition for rules-do-not-require-root
Earlier, a check for "dpkg-build-api (>= 1)" was deciding about it, but
R³ is set to no by default since dpkg version 1.22.13. This change is
hence valid since trixie.

It does not make sense to have silent-on-rules-requires-root since this
is default now, and rules-do-not-require-root should be triggered in all
cases unless d/control says otherwise.