Correctly parse the paragraphs in watch file with format v5
In the previous code, each parsed paragraph began with a newline
("\n"). Therefore, the check for "/^version […]/ at the beginning
fails and the initial paragraph starting with "Version:" is considered
to be a source paragraph. This may result in false positives for
debian-watch-not-mangling-version, for instance.

In the code changed by this commit, the parsed paragraphs end with a
newline, instead.

Closes: #1115463