Priority is set to "optional" by default, even if omitted since dpkg 1.22.12
As mentioned in man/deb-src-control.pod:

Sets the importance of this package in relation to the system as a whole.
The known priorities are B<required>, B<important>, B<standard>, B<optional>,
B<extra>, and B<unknown>, but other values can be used as well.

When omitted,
the I<priority> defaults to B<optional>
(since dpkg 1.22.13).