Correctly parse the paragraphs in watch file with format v5
In the previous code, each parsed paragraph began with a newline
("\n"). Therefore, the check for "/^version […]/ at the beginning
fails and the initial paragraph starting with "Version:" is considered
to be a source paragraph. This may result in false positives for
debian-watch-not-mangling-version, for instance.

In the code changed by this commit, the parsed paragraphs end with a
newline, instead.

I did a first attempt at fixing the issue in MR !619. It did not pass
the Perl::Critic check, because a empty string (`""`) was introduced in
the code. A quote string (`q{}`) is used now.

Closes: #1115463