recursive-privilege-change: do not trigger with -maxdepth 1
maintainer-script-bad-command run a number of tests on the debian
scripts. One of tests is to detect a recursive chmod/chown.

It also checks these in combination with find, but it fails to take into
account if find is used with `-maxdepth 1`. This remedies it and will
still trigger if `-maxdepth N` is used with N > 1