Do not trigger mismatch-changelog in Ubuntu for -proposed
Packages in Ubuntu are usually uploaded to foo-proposed but their
changelog only mentions foo, by convention.

I suspect most Ubuntu developers do not use the sbuild unshare backend
yet (which has a setup helper that enables the pocket *properly* if the
distribution is foo-proposed, but this is bound to change. I personally
just trained my brain to skip those lines, until this actually bit me
due to a mistyped changelog.