Remove library-not-linked-against-libc check
This check made assumptions that are no longer true.

-Wl,--as-needed is set by default since Debian 11
which results in many library files (particularly plugins)
not linking directly against libc.

Ubuntu enables LTO optimization by default and many
Debian packages opt in to that, which also
can trigger this Lintian check. This was seen
with evolution-data-server 3.56.2-5.

Further, the default Salsa CI pipeline fails for
any Lintian error, so we should be cautious
about having Lintian errors.

https://udd.debian.org/lintian-tag/library-not-linked-against-libc?affected=yes
reports that this is emitted 5000 times before getting past the openkim-models
package in an alphabetical list

Closes: #896012
Closes: #909267